+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K97GV6JM2QYXAPVF223QJSJY"
+        "01K997KMD1JNMEK3ARF4SHFM57"
     )
     r"""The unique identifier of the data source"""