+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-11-04T13:27:54.840Z")
+    created: Optional[datetime] = parse_datetime("2025-11-05T05:24:58.805Z")
     r"""The date and time the resource was created"""