+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K97GV6JHKNZSC3D5SWSZ2SRQ"
+        "01K997KMCWEG4Z050ET6WN63SK"
     )
     r"""The unique identifier of the data source"""