+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -80,5 +80,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-04T13:27:51.119Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-05T05:24:55.173Z")
     r"""The date and time the resource was last updated"""