+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K97GV6JKQP6YQQQPKAHCFYDC"
+        "01K997KMCZ26320RAZD0ZEA1NE"
     )
     r"""The unique identifier of the data source"""