+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K97GV6JPTR9YM58CQ20JTSPR"
+        "01K997KMD5SBH9GEKDKHJSF125"
     )
     r"""The unique identifier of the data source"""