+++ packages/orq-rc/RELEASES.md
@@ -3829,3 +3829,13 @@ Based on:
 - [python v4.0.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.7 - packages/orq-rc
+
+## 2025-11-05 05:26:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.648.1 (2.739.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.8 - packages/orq-rc